Enable mass upload for former_students
