+++ app/policies/user_policy.rb
@@ -15,4 +15,16 @@ class UserPolicy < ApplicationPolicy
   def new_with_former_student?
     @user.try :is_admin_or_super_admin?
   end
+
+  def former_students?
+    @user.try :is_admin_or_super_admin?
+  end
+
+  def former_students_upload?
+    @user.try :is_admin_or_super_admin?
+  end
+
+  def import_former_students?
+    @user.try :is_admin_or_super_admin?
+  end
 end
