+++ app/views/students/_former_students_table.html.erb
@@ -1,0 +1,33 @@
+<div id="man" class="col s12 m4 l10">
+    <div class="card material-table">
+    <div class="table-header">
+        <span class="table-title">Exbecarios</span>
+        <div class="actions">
+            <%= link_to 'Registrar', new_with_former_student_path, :class=>"btn-small"%>
+            <%= link_to 'Carga Masiva', former_students_upload_path, :class=>"btn-small"%>
+        </div>
+    </div>
+    <table id="datatable">
+        <thead>
+        <tr>
+            <th>Cvu</th>
+            <th>Nombre</th>
+            <th>  </th>
+        </tr>
+        </thead>
+        <tbody>
+        <% users.each do |user| %>
+            <tr>
+            <td><%= user.student.cvu %></td>
+            <td><%= user.student.full_name %></td>
+            <td><%= link_to 'Show', student_path(user.student.id), :style=>'color:white;font-size:10px;width:100px ! important',:class=>"waves-effect waves-light btn"%><br></br>
+                <% if user.student.curriculum %>
+                    <%= link_to 'Curriculum', curriculum_path(user.student.curriculum), :style=>'color:white;font-size:10px;width:100px ! important',:class=>"waves-effect waves-light btn"%><br></br>
+                <% end %>
+                <%= link_to "Eliminar", user.student, method: :delete, data: { confirm: '¿Estás seguro que quieres eliminar al usuario?' },:style=>'color:white;;font-size:10px;width:100px ! important',:class=>"waves-effect waves-light btn" %></td>
+            </tr>
+        <% end %>
+        </tbody>
+    </table>
+    </div>
+</div>
