+++ app/views/students/former_students.html.erb
@@ -1,0 +1,10 @@
+<main style="padding-left: 240px">
+<div class="row">
+  <div class="col s12"><p></p></div>
+  <div class="col s12 m4 l1"><p></p></div>
+  <%= render "former_students_table", users: @users %>
+  <div class="col s12 m4 l1"><p></p></div>
+</div>
+
+
+</main>
